move promptbuilder and tests
